verificando se existe o recurso antes de deletar

+++ src/main/java/com/acpdq/dscommerce/services/ProductService.java
@@ -55,6 +55,9 @@ public class ProductService {
 	
 	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
+		if( !repository.existsById(id) )
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		
 		try {
 			repository.deleteById(id);
 		} catch (EmptyResultDataAccessException e) {
